Two fixups to reduce potential memory leaks
Using 'valgrind', a potential memory leak in the HistPlotter class was
detected when destructed. A release function for the canvas was added to
let ROOT take care of the memory at the end of the programme.

Two lines were added to reset canvas and legend in MatchLONLO to avoid
ROOT's internal warning message of a canvas being recreated.