Bugfix in MatchLONLO: matrices were not saved
Due to a bug, when using the constructor of MatrixHolderContainer to
pull matrices from a file, their member variable that holds their name
was not set. Because of that, the output files were not saved. This is
now fixed. (The bug did not occur when using the pullHistograms()
function instead of the constructor).