Restructure 4: new Makefile and build structure
The Makefile was completely rewritten. Adding new classes and headers
should work automatically now, so no manual addition of files should be
needed within the Makefile.

Building of the object files (.o) is now moved to a separate directory
'build'. The binary, namely 'plot' is created in the 'bin' directory.
The main.cxx file has been moved to the source as well.