Added get/setFileName() to THolder objects
The THolder (and therefore all inheriting classes) now contain
additional functionality to retrieve the name of the file where the
histograms were pulled from, i.e. getName() will return the title of the
histgoram itself getFileName() will return the name of the file. This is
also added to the pullHistograms() function and the constructor of the
HistHolderContainer and MatrixHolderContainer objects.

The study classes making use of the old functionality were updated.