Optimisation 2: other classes
A couple of changes in style (to be more consistent). Please refer to
individual files for details. Changes were mostly affecting the usage of
initialiser lists with curly brackets, adding 'inline' tags to
functions, removing old bits of code that were not used, and moving
implementations (even of constructors/destructors) to the implementation
files.