Improved plotting of ratios (including arrows)
A function was added to the RatioPlotter class to perform the whole
plotting of the ratios. adjustMarkers() and setOptimalRatioRange() were
removed in the respective classes and functionality was centralised in
the new drawRatio() function.

The function ensures that a maximum of 25% of data points is out of the
ratio range and makes adjustments if necessary. For points out of range,
arrows are plotted to indicate this.