Introduce OptionReader to evaluate command input
The introduced OptionReader enables the user to directly state the
running mode of the program and the input file (without any interactive
input afterwards). The class is now included in the main.cxx and takes
care of the option handling, including exceptions.

The loadFiles() classes of several studies had to be adjusted in order
to deal with the new (and nicer) std::string obejct for the file name
instead of an old C-like char array.