Change of the file structure
The file structure was greatly changed: .h and implementation files are
not stored at different locations anymore, but always in pairs. The core
classes are moved to a folder called 'core' and further distributed in
subfolders that follow logical reasons (e.g. plotter classes are in a
subdirectory called 'plotters'). The studies classes are in their own
subfolder as well. The Makefile was updated as well to fit the new
structure. Object files (*.o) are now produced within the respective
directories of the source files.

Besides, naming conventions of files changed: no camel case, but lower
case with underscores.