Systematically updated names of the output files
The output files used to have more or less random names. This is now
fixed: they all have an identifier at the beginning of the file name,
e.g. 'comp' if it is a comparison plot. The rest of the file name should
clearly state what the file shows to avoid any confusion.