Added shaded errors to on/off comparisons
In the comparison plots of the on-shell, off-shell and combined
likelihood, the style of the combined LL was changed to use histogram
bars instead of markers. This looks now like a proper reference for the
other two in the background. Instead of error crosses, the bars have
shaded errors now.

To get the correct order of plotting, the order of the input files was
also adjusted:
 - matcher output: combined, on-shell, off-shell
 - likelihood output: combined, on-shell, off-shell